ToppersSlice parseInt chnage into parseFloat

+++ src/requests/ExamResult/RequestToppers.ts
@@ -185,11 +185,11 @@ export const ClassToppersList =
             response.data.GetSelectedSubjectTopperList.map((item, i) => {
                 if (
                     !Subjects.includes(
-                        item.Subject_Name + '#' + item.Rank_Image + '#' + `${parseInt(item.Total_Marks_Scored).toString()} / ${item.Subject_Total_Marks}`
+                        item.Subject_Name + '#' + item.Rank_Image + '#' + `${parseFloat(item.Total_Marks_Scored).toString()} / ${item.Subject_Total_Marks}`
                     )
                 ) {
                     Subjects.push(
-                        item.Subject_Name + '#' + item.Rank_Image + '#' + `${parseInt(item.Total_Marks_Scored).toString()} / ${item.Subject_Total_Marks}`
+                        item.Subject_Name + '#' + item.Rank_Image + '#' + `${parseFloat(item.Total_Marks_Scored).toString()} / ${item.Subject_Total_Marks}`
                     );
                 }
             });
@@ -205,7 +205,7 @@ export const ClassToppersList =
                 response.data.GetSelectedSubjectTopperList.map((item, i) => {
                     if (
                         obj ==
-                        item.Subject_Name + '#' + item.Rank_Image + '#' + `${parseInt(item.Total_Marks_Scored).toString()} / ${item.Subject_Total_Marks}`
+                        item.Subject_Name + '#' + item.Rank_Image + '#' + `${parseFloat(item.Total_Marks_Scored).toString()} / ${item.Subject_Total_Marks}`
                     ) {
                         child.Students.push({
                             Id: item.Student_Id,
